% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_metrics.R
\name{calibrate_metrics}
\alias{calibrate_metrics}
\title{Calculate metrics}
\usage{
calibrate_metrics(object, metric_list)
}
\arguments{
\item{object}{a tibble of class \code{smp_spl_results}}

\item{metric_list}{a list of metrics to evaluate independence of \code{.resid}}
}
\description{
Assumes each row of \code{.resid} list-column is a tibble with also contains a
column \code{.resid}.
}
