% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_splits-model_spec.R,
%   R/fit_splits-workflow.R, R/fit_splits.R
\name{fit_splits.model_spec}
\alias{fit_splits.model_spec}
\alias{fit_splits.workflow}
\alias{fit_splits}
\title{Fit multiple sample splits}
\usage{
\method{fit_splits}{model_spec}(
  object,
  formula,
  data,
  num_analysis,
  num_assessment,
  ...,
  metrics = NULL,
  control = tune::control_resamples()
)

\method{fit_splits}{workflow}(
  object,
  data,
  num_analysis,
  num_assessment,
  ...,
  metrics = NULL,
  control = tune::control_resamples()
)

fit_splits(object, ...)
}
\arguments{
\item{object}{A \code{parsnip} model specification or a \code{\link[workflows:workflow]{workflows::workflow()}}.
No tuning parameters are allowed.}

\item{formula}{An object of class \code{formula} (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame of predictors and outcomes to use when fitting the
workflow.}

\item{num_analysis}{A positive integer specifying number of observations for
analysis.}

\item{num_assessment}{A positive integer specifying number of observations
for goodness-of-fit tests.}

\item{...}{Currently unused.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}, or \code{NULL} to compute a standard
set of metrics.}

\item{control}{A \code{\link[tune:control_resamples]{control_resamples()}} object used to fine tune the resampling
process.}
}
\description{
Fits model according to sample splitting on a grid of values determined by
\code{num_analysis} and \code{num_assessment}, where each tuple determines the number
of analysis and assessment sets.
}
