% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_splits-fable.R
\name{fit_splits_tbl_ts}
\alias{fit_splits_tbl_ts}
\title{Fit multiple sample splits to fable models}
\usage{
fit_splits_tbl_ts(.data, .model, num_analysis, num_assessment, ...)
}
\arguments{
\item{.data}{A data structure suitable for the models (such as a \code{tsibble}).}

\item{.model}{Definition for the model to be used.}

\item{num_analysis}{A positive integer specifying number of observations for
analysis.}

\item{num_assessment}{A positive integer specifying number of observations
for goodness-of-fit tests.}

\item{...}{Further arguments passed to methods.}
}
\description{
Fit multiple sample splits to fable models
}
\examples{
library(fable)
data <-
     tsibble::tsibble(value = rnorm(100), date = Sys.Date() + 0:99, index = date)
# Consider an ARMA(1, 1) model
#data |>
#     fit_splits(model_dfn = ARIMA(value ~ pdq(1, 0, 1)), 2, 2)
}
