% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_splits-list.R
\name{fit_splits_list}
\alias{fit_splits_list}
\title{Fit multiple sample splits from a list of models}
\usage{
fit_splits_list(
  object,
  data,
  num_analysis,
  num_assessment,
  use_fit_resamples = FALSE,
  ...,
  metrics = NULL,
  control = tune::control_resamples()
)
}
\arguments{
\item{object}{list of \code{model_spec} and/or \code{workflow} object to be fit}

\item{data}{A data frame of predictors and outcomes to use when fitting the
workflow.}

\item{num_analysis}{A positive integer specifying number of observations for
analysis.}

\item{num_assessment}{A positive integer specifying number of observations
for goodness-of-fit tests.}

\item{use_fit_resamples}{If \code{TRUE}, uses \code{\link[tune:fit_resamples]{tune::fit_resamples()}}.}

\item{...}{Currently unused.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}, or \code{NULL} to compute a standard
set of metrics.}

\item{control}{A \code{\link[tune:control_resamples]{control_resamples()}} object used to fine tune the resampling
process.}
}
\description{
Fit multiple sample splits from a list of models
}
