% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_boot_splits-model_spec.R,
%   R/par_boot_splits-workflow.R, R/par_boot_splits.R
\name{par_boot_splits.model_spec}
\alias{par_boot_splits.model_spec}
\alias{par_boot_splits.workflow}
\alias{par_boot_splits}
\title{Parametric bootstrap quantiles}
\usage{
\method{par_boot_splits}{model_spec}(
  object,
  formula,
  data,
  num_analysis,
  num_resamples,
  resample_size,
  quantiles,
  burn_in = 200L,
  ...,
  metrics,
  control
)

\method{par_boot_splits}{workflow}(
  object,
  data,
  num_analysis,
  num_resamples,
  resample_size,
  quantiles,
  burn_in = 200L,
  ...,
  metrics,
  control
)

par_boot_splits(object, ...)
}
\arguments{
\item{object}{either of class \code{model_spec} or \code{workflow}. The fitted object
should have \code{simts::simts} method implemented.}

\item{formula}{An object of class \code{formula} (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame of predictors and outcomes to use when fitting the
workflow.}

\item{num_analysis}{A positive integer specifying number of observations for
analysis.}

\item{num_resamples}{positive integer. Number of bootstrap resamples.}

\item{resample_size}{positive integer. Number of residuals per bootstrap
resample.}

\item{quantiles}{vector of probabalities at which quantiles of metrics are
calculated.}

\item{burn_in}{positive integer. Number of residuals used for burn-in.}

\item{...}{Currently unused.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}, or \code{NULL} to compute a standard
set of metrics.}

\item{control}{A \code{\link[tune:control_resamples]{control_resamples()}} object used to fine tune the resampling
process.}
}
\description{
Yields parametric bootstrapped metrics for specified sample splits.
}
