% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpspl-package.R
\docType{package}
\name{smpspl-package}
\alias{smpspl-package}
\alias{_PACKAGE}
\title{smpspl: Sample Splitting}
\description{
Provides functions that perform sample splitting as described in upcoming manuscript.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leonfernandes/smpspl}
  \item Report bugs at \url{https://github.com/leonfernandes/smpspl/issues}
}

}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.edu} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

}
\keyword{internal}
