% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpspl.R
\name{smpspl_boot}
\alias{smpspl_boot}
\title{Bootstrapped Sample Split Residuals}
\usage{
smpspl_boot(
  object,
  data,
  f_n,
  l_n,
  num_resamples,
  resample_size,
  burn_in = 200L,
  ...
)
}
\arguments{
\item{object}{An object to fit a time series model to. Implemented models are
\code{model_spec}, \code{workflow} or \code{mdl_defn} objects.}

\item{data}{A \link[tsibble:tsibble-package]{tsibble} of predictors and outcomes
to use when fitting the model.}

\item{f_n}{positive integer. Number of terms in analysis split.}

\item{l_n}{positive integer. Number of terms in assessment split.}

\item{num_resamples}{positive integer. Number of bootstrap resamples.}

\item{resample_size}{positive integer. Number of residuals per bootstrap
resample.}

\item{burn_in}{positive integer. Number of residuals used for burn-in.}

\item{...}{Additional arguments needed to fit the data such as formula.}
}
\description{
Bootstrapped Sample Split Residuals
}
