% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpspl_table.R
\name{smpspl_table}
\alias{smpspl_table}
\title{Sample Split Residuals for a List of Models}
\usage{
smpspl_table(model_list, data, f_n, l_n, ...)
}
\arguments{
\item{model_list}{list of models to be fit}

\item{data}{A \link[tsibble:tsibble-package]{tsibble} of predictors and outcomes
to use when fitting the model.}

\item{f_n}{positive integer. Number of terms in analysis split.}

\item{l_n}{positive integer. Number of terms in assessment split.}

\item{...}{Additional arguments needed to fit the data such as formula.}
}
\description{
Sample Split Residuals for a List of Models
}
