% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpspl_table.R
\name{smpspl_table_grid}
\alias{smpspl_table_grid}
\title{Sample Split Residuals on a Grid for a List of Models}
\usage{
smpspl_table_grid(model_list, data, num_analysis, num_assessment, ...)
}
\arguments{
\item{model_list}{list of models to be fit}

\item{data}{A \link[tsibble:tsibble-package]{tsibble} of predictors and outcomes
to use when fitting the model.}

\item{num_analysis}{A positive integer specifying number of observations for
analysis.}

\item{num_assessment}{A positive integer specifying number of observations
for goodness-of-fit tests.}

\item{...}{Additional arguments needed to fit the data such as formula.}
}
\description{
Sample Split Residuals on a Grid for a List of Models
}
