% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_resids.R
\name{subset_resids}
\alias{subset_resids}
\title{Subsets tail of residuals}
\usage{
subset_resids(object, resid_col, size_col, metrics = NULL)
}
\arguments{
\item{object}{a tibble.}

\item{resid_col}{single character string. Column name corresponding to
residuals.}

\item{size_col}{single character string. Column name corresponding to number
of residuals to be subset from the tail.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}}, or \code{NULL} to compute a standard
set of metrics.}
}
\description{
Helper function to subset a list-column of residuals \code{resid_col} based on
\code{size_col}. The subset column will have name \code{.resid}; note that this
overwrites any exisiting column of the same name.
}
