% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_metrics.R
\name{tune_metrics}
\alias{tune_metrics}
\title{Calculate metrics}
\usage{
tune_metrics(object, metric_list)
}
\arguments{
\item{object}{a tibble of class \code{smp_spl_results}}

\item{metric_list}{a list of metrics to be evaluate \code{.resids}}
}
\description{
Calculate metrics
}
