% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_splits.R
\name{sample_splits}
\alias{sample_splits}
\title{Creates sample splits}
\usage{
sample_splits(n, num_analysis, num_assessment)
}
\arguments{
\item{n}{Number of terms to sample split.}

\item{num_analysis}{A positive integer specifying number of observations for
analysis.}

\item{num_assessment}{A positive integer specifying number of observations
for goodness-of-fit tests.}
}
\value{
An object of class \code{grid_tbl}.
}
\description{
Fits model according to sample splitting on a grid of values determined by
\code{num_analysis} and \code{num_assessment}, where each tuple determines the number
of analysis and assessment sets.
}
\examples{
library(fable)
library(smpspltools)
data <-
     tsibble::tsibble(x = rnorm(100), date = Sys.Date() + 0:99, index = date)
# Consider an AR(1) model
o <-
     ARIMA(x ~ pdq(1, 0, 0) + PDQ(0, 0, 0)) |>
     smpspl_grid(data, 2, 2)
}
