% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_splits.R
\name{sample_splits}
\alias{sample_splits}
\title{Creates sample splits}
\usage{
sample_splits(data, num_analysis, num_assessment)
}
\arguments{
\item{data}{A data frame of predictors and outcomes.}

\item{num_analysis}{positive integer specifying number of observations for
analysis.}

\item{num_assessment}{positive integer specifying number of observations for
goodness-of-fit tests.}
}
\value{
An object of class \code{smp_spl}.
}
\description{
Creates sample splits
}
