% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpspl.R
\name{smpspl}
\alias{smpspl}
\title{Sample Split Residuals}
\usage{
smpspl(object, data, f_n, l_n, ...)
}
\arguments{
\item{object}{An object to fit a time series model to. Implemented models are
\code{model_spec}, \code{workflow} or \code{mdl_defn} objects.}

\item{data}{A \link[tsibble:tsibble-package]{tsibble} of predictors and outcomes
to use when fitting the model.}

\item{f_n}{positive integer. Number of terms in analysis split.}

\item{l_n}{positive integer. Number of terms in assessment split.}

\item{...}{Additional arguments needed to fit the data such as formula.}
}
\value{
a \link[tsibble:tsibble-package]{tsibble} of class "smpspl".
}
\description{
Evaluates residuals of a model for a single sample split.
}
\examples{
library(fable)
library(smpspltools)
library(generics)
data <-
     tsibble::tsibble(x = rnorm(100), date = Sys.Date() + 0:99, index = date)
# Consider an AR(1) model
o <-
     ARIMA(x ~ pdq(1, 0, 0) + PDQ(0, 0, 0)) |>
     smpspl(data, 50, 100)
# Calculate lanyard features
my_acf <-
     function(x) {
         data.frame(t = x, e = 0) |>
             lanyard::acf_metric(t, e) |>
             tidy()
     }
o |>
     features(.resid, features = my_acf)
library(parsnip)
library(modeltime)
# Repeat with `modeltime::arima_reg`
arima_spec <-
     arima_reg() |>
     set_engine("arima")
o2 <- smpspl(arima_spec, data, 50, 100, formula = x ~ date)
o2 |>
     features(.resid, features = my_acf)
}
