% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpspl_boot.R
\name{smpspl_boot}
\alias{smpspl_boot}
\title{Bootstrapped Sample Split Residuals}
\usage{
smpspl_boot(object, data, f_n, l_n, num_resamples, ...)
}
\arguments{
\item{object}{An object to fit a time series model to. Implemented models are
\code{model_spec}, \code{workflow} or \code{mdl_defn} objects.}

\item{data}{A \link[tsibble:tsibble-package]{tsibble} of predictors and outcomes
to use when fitting the model.}

\item{f_n}{positive integer. Number of terms in analysis split.}

\item{l_n}{positive integer. Number of terms in assessment split.}

\item{num_resamples}{positive integer. Number of bootstrap resamples.}

\item{...}{Additional arguments needed to fit the data such as formula.}
}
\value{
a \link[tsibble:tsibble-package]{tsibble} of bootstrap resampled
residuals.
}
\description{
Bootstrapped Sample Split Residuals
}
\examples{
library(fable)
data <-
     tsibble::tsibble(x = rnorm(100), date = Sys.Date() + 0:99, index = date)
# Consider an AR(1) model
o <-
     ARIMA(x ~ pdq(1, 0, 0) + PDQ(0, 0, 0)) |>
     smpspl_boot(data, 50, 100, 20)
}
